+++ test/rails_app/app/controllers/admin/rodauth_controller.rb
@@ -1,0 +1,2 @@
+class Admin::RodauthController < ApplicationController
+end