+++ test/rails_app/app/lib/rodauth_admin.rb
@@ -1,5 +1,6 @@
 class RodauthAdmin < Rodauth::Rails::Auth
   configure do
     prefix "/admin"
+    rails_controller { Admin::RodauthController }
   end
 end